// Code/Websites/DanpheEMR/wwwroot/DanpheApp/src/app/ot/shared/dto/get-ot-anaesthesia.dto.ts
export class GetOTAnaesthesia_DTO {
    AnaesthesiaId: number = 0;
    ServiceItemId: number = null;
    ServiceDepartmentId: number = null;
    IntegrationItemId: number = null;
    IntegrationName: string = null;
    ItemCode: string = null;
    ItemName: string = "";
    IsTaxApplicable: boolean;
    Description: string = null;
    DisplaySeq: number = null;
    IsDoctorMandatory: boolean;
    IsOT: boolean;
    IsProc: boolean;
    ServiceCategoryId: number = null;
    AllowMultipleQty: boolean;
    DefaultDoctorList: string = null;
    IsValidForReporting: boolean;
    IsErLabApplicable: boolean;
    CreatedBy: number = null;
    CreatedOn: string = null;
    ModifiedOn: string = null;
    ModifiedBy: number = null;
    IsActive: boolean;
    IsIncentiveApplicable: boolean;
    ServiceDepartmentName: string = null;
    AnaesthesiaTypeId: number = null;
    Price: number = null;
}