// Code/Websites/DanpheEMR/wwwroot/DanpheApp/src/app/ot/shared/dto/get-ot-personnel.dto.ts
export class GetOTPersonnel_DTO {
    EmployeeId: number = null;
    Salutation: string = null;
    FirstName: string = null;
    MiddleName: string = null;
    LastName: string = null;
    FullName: string = "";
    DateOfBirth: string = null;
    DateOfJoining: string = null;
    ContactNumber: string = null;
    ContactAddress: string = null;
    Email: string = null;
    Gender: string = null;
    Extension: string = null;
    SpeedDial: string = null;
    OfficeHour: string = null;
    RoomNo: string = null;
    IsActive: boolean = null;
    MedCertificationNo: string = null;
    Signature: string = null;
    LongSignature: string = null;
    DepartmentId: number = null;
    DepartmentName: string = null;
    EmployeeRoleId: number = null;
    EmployeeRoleName: string = null;
    EmployeeTypeId: number = null;
    EmployeeTypeName: string = null;
    IsAppointmentApplicable: boolean = null;
    LabSignature: string = null;
    CreatedOn: string = null;
    CreatedBy: string = null;
    SignatoryImageName: string = null;
    DisplaySequence: number = null;
    TDSPercent: number = null;
    PANNumber: string = null;
    IsIncentiveApplicable: boolean = null;
    RadiologySignature: string = null;
    BloodGroup: string = null;
    NursingCertificationNo: string = null;
    HealthProfessionalCertificationNo: string = null;
    DriverLicenseNo: string = null;
    OpdNewPatientServiceItemId: number = null;
    FollowupServiceItemId: number = null;
    OpdOldPatientServiceItemId: number = null;
    InternalReferralServiceItemId: number = null;
    PersonnelTypeId: number = null;
    PersonnelTypeName: string = null;
}