// Code/Websites/DanpheEMR/wwwroot/DanpheApp/src/app/ot/shared/dto/get-ot-team-info.dto.ts
export class GetOTTeamInfo_DTO {
    TeamInfoId: number = null;
    PersonnelTypeId: number = null;
    PersonnelType: string = null;
    EmployeeId: number = null;
    FullName: string = null;
    OTBookingId: number = null;
    PatientId: number = null;
    PatientVisitId: number = null;
    CreatedBy: number = null;
    CreatedOn: string = null;
    ModifiedBy: number = null;
    ModifiedOn: string = null;
}